Extract chat balloons into UI::Chat
Move the chat-balloon subsystem (the CHAT message buffer, CreateChat/AssignChat/
MoveChat/AddChat/AddGuildName, the bubble layout/render in SetBooleanPosition/
RenderBoolean/RenderBooleans, and SetPlayerColor, ~620 lines) out of
ZzzInterface.cpp into UI/Chat/Chat.{h,cpp} under namespace UI::Chat, alongside
UI::Chat::Whisper. Callers across six files are now qualified. The RenderDebugText
global that sat in the moved range stays put. No behaviour change.
